Updated the  file type
