+++ mobile-app/app/(tabs)/academic/kuppi.jsx
