+++ mobile-app/app/(tabs)/academic/resources.jsx
